Tweak/player shooting rework

+++ Assets/2_MainScene/S2_Scripts/S2_BossScripts/S2_BossStats.cs
@@ -10,11 +10,11 @@ public class S2_BossStats : ScriptableObject
     [SerializeField] protected float maxSpeed;
     [SerializeField] protected float minSpeed;
     [SerializeField] protected float[] attributes; //{ FirePower, RapidFireSpeed, Shields, Health }
-    [SerializeField] protected float points;
+    [SerializeField] protected int points;
     public float GetMaxSpeed() { return maxSpeed; }
 
     public float[] GetAttributes() { return attributes; }
 
     public float GetMinSpeed() { return minSpeed; }
-    public float GetPointValue() { return points; }
+    public int GetPointValue() { return points; }
 }
